refactor folder structure following new conventionsg
