+++ lib/features/new_posts/widgets/header_input_text.dart
