+++ lib/features/new_posts/widgets/input_text.dart
