+++ lib/features/posts/post_view.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:intl/intl.dart';
 
-import '../../../flavors.dart';
-import '../../../ui/refresh_list_adaptive.dart';
-import '../view_model/post_view_model.dart';
-import '../widgets/post_card_item.dart';
+import '../../design_system/widgets/refresh_list_adaptive.dart';
+import '../../flavors.dart';
+import 'post_view_model.dart';
+import 'widgets/post_card_item.dart';
 
 class PostView extends StatelessWidget {
   const PostView({
@@ -23,8 +23,7 @@ class PostView extends StatelessWidget {
         actions: <Widget>[
           IconButton(
             icon: const Icon(Icons.add),
-            onPressed: () =>
-                postViewModel.openNewPostBottomSheet(context, postViewModel),
+            onPressed: () => postViewModel.openNewPostBottomSheet(context, postViewModel),
           )
         ],
       ),
@@ -38,8 +37,7 @@ class PostView extends StatelessWidget {
                 key: Key(post.id.toString()),
                 id: post.id.toString(),
                 text: post.text,
-                createdAt:
-                    DateFormat.yMd().add_Hms().format(post.createdAtDatetime),
+                createdAt: DateFormat.yMd().add_Hms().format(post.createdAtDatetime),
               );
             },
             itemCount: postViewModel.posts.length,
