+++ test/features/posts/view_model/post_view_model_test.dart
@@ -1,5 +1,5 @@
+import 'package:flutter_mobx_template/features/posts/post_view_model.dart';
 import 'package:flutter_mobx_template/models/post.dart';
-import 'package:flutter_mobx_template/modules/posts/view_model/post_view_model.dart';
 import 'package:flutter_mobx_template/repository/post_repository.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
