+++ test/fixtures/fixture.dart
@@ -1,5 +1,4 @@
 import 'dart:convert';
 import 'dart:io';
 
-dynamic fixture(String filename) =>
-    jsonDecode(File('./test/fixtures/$filename').readAsStringSync());
+dynamic fixture(String filename) => jsonDecode(File('./test/fixtures/$filename').readAsStringSync());
